Renamed and replaced relevant files
Replaced outdated files with latest files, added "Screen" to the end of each files name to follow the convention.
