+++ frontend/HarmonyMobileApp/Components/navigation.component.js
@@ -21,14 +21,11 @@ import PairingResultsScreen from "../screens/PairingResultsScreen.js";
 import DrinkDetailsScreen from "../screens/DrinkDetailsScreen.js";
 
 import config from "../aws-exports";
-// import SignIn from "../screens/SignInScreen";
-import SignIn from "../screens/SignIn";
-// import SignUp from "../screens/SignUpScreen";
-import SignUp from "../screens/SignUp";
-// import ConfirmSignUp from "../screens/ConfirmSignUpScreen";
-import ConfirmSignUp from "../screens/ConfirmSignUp";
-import ForgotPassword from "../screens/ForgotPassword";
-import ConfirmForgotPassword from "../screens/ConfirmForgotPassword";
+import SignIn from "../screens/SignInScreen";
+import SignUp from "../screens/SignUpScreen";
+import ConfirmSignUp from "../screens/ConfirmSignUpScreen";
+import ForgotPassword from "../screens/ForgotPasswordScreen";
+import ConfirmForgotPassword from "../screens/ConfirmForgotPasswordScreen";
 
 Amplify.configure(config);
